feature/86cw4nf56: Created the bookings helper function
