+++ behind-the-veil-siteroot/imports/ui/components/pages/profile/BrideProfileTabs.jsx
@@ -1,21 +1,17 @@
 /**
  * File Description: Brides profile page tabs
- * File version: 2.3
+ * File version: 2.4
  * Contributors: Katie, Nikki
  */
 
 import React from 'react';
-import {useSubscribe, useTracker} from "meteor/react-meteor-data";
-
-import BookingCollection from "../../../../api/collections/bookings";
-import ServiceCollection from "../../../../api/collections/services";
-import ImageCollection from "../../../../api/collections/images";
 
 import Tabs from "../../../components/tabs/Tabs";
 import Loader from "../../../components/loader/Loader";
 import BookingListView from "../../../components/booking/BookingListView";
 import BookingCard from "../../../components/card/BookingCard";
 import BookingStatus from "../../../../ui/enums/BookingStatus";
+import {getBookings} from "../../DatabaseHelper";
 
 /**
  * Component for bride's profile tabs
@@ -25,51 +21,14 @@ import BookingStatus from "../../../../ui/enums/BookingStatus";
 export const BrideProfileTabs = ({userInfo}) => {
 
     // get bookings information from database
-    const isLoadingBooking = useSubscribe('all_user_bookings', userInfo.username);
-    const isLoadingService = useSubscribe('all_services');
-    const isLoadingServiceImage = useSubscribe('service_images');
-
-    let bookingsData = useTracker(() => {
-
-        return BookingCollection.find({
-            $or: [
-                {"brideUsername": userInfo.username},
-                {"artistUsername": userInfo.username}
-            ]
-        }).fetch();
-    });
-
-    let servicesData = useTracker(() => {
-        return ServiceCollection.find().fetch();
-    });
-
-    let imagesData = useTracker(() => {
-        return ImageCollection.find({"imageType": "service"}).fetch();
-    });
-
-    const isLoading = isLoadingBooking() || isLoadingService() || isLoadingServiceImage();
-
-    // manual aggregation into bookingsData with its services and images
-    for (let i = 0; i < bookingsData.length; i++) {
-
-        // aggregate with service first
-        for (let j = 0; j < servicesData.length; j++) {
-            // find matching service ID
-            if (bookingsData[i].serviceId === servicesData[j]._id) {
-                bookingsData[i].serviceName = servicesData[j].serviceName;
-                bookingsData[i].serviceDesc = servicesData[j].serviceDesc;
-                break;
-            }
-        }
-        // then aggregate with the FIRST service image (cover)
-        for (let j = 0; j < imagesData.length; j++) {
-            // find matching image for the service
-            if (imagesData[j].imageType === "service" && bookingsData[i].serviceId === imagesData[j].target_id) {
-                bookingsData[i].serviceImageData = imagesData[j].imageData;
-                break;
-            }
-        }
+    // get bookings information from database
+    const bookingFilter = {
+        $or: [
+            { "brideUsername": userInfo.username },
+            { "artistUsername": userInfo.username }
+        ]
     }
+    const [isLoading, bookingsData] = getBookings("all_user_bookings", [userInfo.username], bookingFilter);
 
     // wait for bookings data to be loaded
     if (isLoading) {
