+++ behind-the-veil-siteroot/imports/ui/components/pages/service/AddEditServicePage.jsx
@@ -17,7 +17,6 @@ import Input from "../../input/Input";
 import Button from "../../button/Button";
 
 import { CheckIcon, XMarkIcon, TrashIcon } from "@heroicons/react/24/outline";
-import UrlBasePath from "../../../enums/UrlBasePath";
 
 export const AddEditServicePage = ({ isEdit }) => {
     const navigateTo = useNavigate();
@@ -31,7 +30,8 @@ export const AddEditServicePage = ({ isEdit }) => {
     const [servicePrice, setServicePrice] = useState(0);
     const [serviceDescription, setServiceDescription] = useState("");
 
-    userInfo = getUserInfo();
+    // check user owns service + user is an artist, else redirect to home
+    const userInfo = getUserInfo();
     if (userInfo.type !== "artist") {
         navigateTo(`/`);
     }
@@ -63,9 +63,6 @@ export const AddEditServicePage = ({ isEdit }) => {
             });
         }
     }, []);
-
-    // check user owns service + user is a artist, else redirect to home
-    // after rebase: const userInfo = getUserData();
 
     const [filesArray, setFilesArray] = useState([]);
     const [fileRejected, setFileRejected] = useState(false);
